requirements.txt
absl-py==1.2.0
asgiref==3.5.2
Babel==2.10.3
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
crispy-bootstrap5==0.6
Django==4.0.6
django-crispy-forms==1.14.0
django-jazzmin==2.5.0
django-phonenumber-field==6.3.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.48.0
idna==3.3
importlib-metadata==4.12.0
Markdown==3.4.1
MarkupSafe==2.1.1
more-itertools==8.13.0
mysqlclient==2.1.1
numpy==1.23.1
oauthlib==3.2.0
phonenumbers==8.12.52
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-decouple==3.6
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
urllib3==1.26.11
Werkzeug==2.2.1
zipp==3.8.1
